+++ src/main/java/com/musseukpeople/woorimap/post/domain/image/PostImage.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.post.domain;
+package com.musseukpeople.woorimap.post.domain.image;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -8,6 +8,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+
+import com.musseukpeople.woorimap.post.domain.Post;
 
 import lombok.AccessLevel;
 import lombok.Getter;
